libavformat/matroska_ids: reorder defines to match a deterministic order
The order is the alphabetical order of the element path.

We can't sort by IDs as they are not related to the path of the element nor
have incremental versions.

In the stable spec the path and element names should be stable.

It allows a logical positioning of child elements after their parent element.
For example MATROSKA_ID_BLOCKADDITIONS was mistakenly positioned as a child of
Cluster.

No functional change. The defines have just been moved around.