avformat/matroskaenc: add an option to set more accurate timestamps
By default the precision of Block timestamps is 1 ms. Using the
TrackTimestampScale it's possible to set these timestamp at almost exactly
their fractional value. The rounding the values will always end up on the
proper clock tick for that track.

Players will need to be aware of the TrackTimestampScale to be able to read
the file properly as it influences the timestamp values stored in each Block.

Each Block timestamp "tick" corresponds to a tick in the track clock. For
audio with high frequency it's recommended to pack a lot of samples in a frame
otherwise the possible duration of a Cluster is very small :
  (65536/frequency) seconds
A factor of 8 should work for most cases but finer values should be used per
codec.

This should not be used for tracks without a fixed clocked.

This mode is not supported in WebM as it doesn't support TrackTimestampScale.